requirements.txt
pandas
google-generativeai
Pillow
dotenv
os